Move date and time display to header
